requirements.txt
beautifulsoup4~=4.12.0
requests~=2.28.2
selenium==4.8.3
python-dotenv